Put Enums in their own folder so that they don't get mocked.
