+++ backend/entity/MonsterEnums.ts
@@ -1,0 +1,60 @@
+export enum Size {
+    Tiny = "Tiny",
+    Small = "Small",
+    Medium = "Medium",
+    Large = "Large",
+    Huge = "Huge",
+    Gargantuan = "Gargantuan"
+}
+
+export enum MonsterType {
+    Beast = "Beast",
+    Dragon = "Dragon",
+    Monstrosity = "Monstrosity",
+    Humanoid = "Humanoid",
+    Fiend = "Fiend",
+    Undead = "Undead",
+    Giant = "Giant",
+    Elemental = "Elemental",
+    SwarmOfTinyBeasts = "SwarmOfTinyBeasts",
+    Construct = "Construct",
+    Celestial = "Celestial",
+    Aberration = "Aberration",
+    Fey = "Fey",
+    Plant = "Plant",
+    Ooze = "Ooze"
+}
+
+export enum MonsterRace {
+    AnyRace = "AnyRace",
+    Devil = "Devil",
+    Demon = "Demon",
+    Human = "Human",
+    Shapechanger = "Shapechanger",
+    Goblinoid = "Goblinoid",
+    Titan = "Titan",
+    Gnoll = "Gnoll",
+    Gnome = "Gnome",
+    Dwarf = "Dwarf",
+    Elf = "Elf",
+    Orc = "Orc",
+    Kobold = "Kobold",
+    Lizardfolk = "Lizardfolk",
+    Merfolk = "Merfolk",
+    Sahuagin = "Sahuagin",
+    Grimlock = "Grimlock"
+}
+
+export enum Alignment {
+    Unaligned = "Unaligned",
+    AnyAlignment = "AnyAlignment",
+    LawfulDood = "LawfulDood",
+    LawfulNeutral = "LawfulNeutral",
+    LawfulEvil = "LawfulEvil",
+    NeutralGood = "NeutralGood",
+    Neutral = "Neutral",
+    NeutralEvil = "NeutralEvil",
+    ChaoticGood = "ChaoticGood",
+    ChaoticNeutral = "ChaoticNeutral",
+    ChaoticEvil = "ChaoticEvil"
+}
